Dashboard Refresh & Code Reorganization
A brand new dashboard with user authentication, registration via email confirmation, oAuth, password recovery, and captcha protection, and Soft Deletes for users

+++ database/migrations/2018_04_30_164900_create_articles_tables.php
@@ -13,42 +13,49 @@ class CreateArticlesTables extends Migration
      */
     public function up()
     {
-	    Schema::create('articles', function (Blueprint $table) {
-		    $table->increments('id');
-		    $table->char('iso', 3)->index();
-		    $table->foreign('iso')->references('iso')->on('languages')->onUpdate('cascade')->onDelete('cascade');
-		    $table->integer('organization_id')->unsigned();
-		    $table->foreign('organization_id')->references('id')->on('organizations')->onUpdate('cascade')->onDelete('cascade');
-		    $table->string('user_id', 64);
-		    $table->foreign('user_id')->references('id')->on('users')->onUpdate('cascade')->onDelete('cascade');
-		    $table->string('cover')->nullable();
-		    $table->string('cover_thumbnail')->nullable();
-		    $table->timestamps();
-	    });
 
-	    Schema::create('article_translations', function (Blueprint $table) {
-		    $table->integer('article_id')->unsigned();
-		    $table->foreign('article_id')->references('id')->on('articles')->onUpdate('cascade')->onDelete('cascade');
-		    $table->char('iso',3);
-		    $table->foreign('iso')->references('iso')->on('languages')->onUpdate('cascade');
-		    $table->string('name');
-		    $table->text('description')->nullable();
-		    $table->boolean('vernacular')->default(0);
-		    $table->unique(['article_id','iso'], 'unq_article_translations');
-		    $table->timestamps();
-	    });
+	    if(!Schema::connection('dbp_users')->hasTable('articles')) {
+		    Schema::connection('dbp_users')->create('articles', function (Blueprint $table) {
+			    $table->increments('id');
+			    $table->char('iso', 3)->index();
+			    $table->foreign('iso')->references('iso')->on('dbp.languages')->onUpdate('cascade')->onDelete('cascade');
+			    $table->integer('organization_id')->unsigned();
+			    $table->foreign('organization_id')->references('id')->on('dbp.organizations')->onUpdate('cascade')->onDelete('cascade');
+			    $table->string('user_id', 64);
+			    $table->foreign('user_id')->references('id')->on('users')->onUpdate('cascade')->onDelete('cascade');
+			    $table->string('cover')->nullable();
+			    $table->string('cover_thumbnail')->nullable();
+			    $table->timestamps();
+		    });
+	    }
 
-	    Schema::create('article_tags', function (Blueprint $table) {
-		    $table->integer('article_id')->unsigned();
-		    $table->foreign('article_id')->references('id')->on('articles')->onUpdate('cascade')->onDelete('cascade');
-		    $table->char('iso',3);
-		    $table->foreign('iso')->references('iso')->on('languages')->onUpdate('cascade');
-		    $table->string('tag');
-		    $table->string('name');
-		    $table->text('description')->nullable();
-		    $table->unique(['article_id', 'iso'], 'unq_article_tags');
-		    $table->timestamps();
-	    });
+	    if(!Schema::connection('dbp_users')->hasTable('article_translations')) {
+		    Schema::connection('dbp_users')->create('article_translations', function (Blueprint $table) {
+			    $table->integer('article_id')->unsigned();
+			    $table->foreign('article_id')->references('id')->on('articles')->onUpdate('cascade')->onDelete('cascade');
+			    $table->char('iso', 3);
+			    $table->foreign('iso')->references('iso')->on('dbp.languages')->onUpdate('cascade');
+			    $table->string('name');
+			    $table->text('description')->nullable();
+			    $table->boolean('vernacular')->default(0);
+			    $table->unique(['article_id', 'iso'], 'unq_article_translations');
+			    $table->timestamps();
+		    });
+	    }
+
+	    if(!Schema::connection('dbp_users')->hasTable('article_tags')) {
+		    Schema::connection('dbp_users')->create('article_tags', function (Blueprint $table) {
+			    $table->integer('article_id')->unsigned();
+			    $table->foreign('article_id')->references('id')->on('articles')->onUpdate('cascade')->onDelete('cascade');
+			    $table->char('iso', 3);
+			    $table->foreign('iso')->references('iso')->on('dbp.languages')->onUpdate('cascade');
+			    $table->string('tag');
+			    $table->string('name');
+			    $table->text('description')->nullable();
+			    $table->unique(['article_id', 'iso'], 'unq_article_tags');
+			    $table->timestamps();
+		    });
+	    }
 
     }
 
@@ -59,8 +66,8 @@ class CreateArticlesTables extends Migration
      */
     public function down()
     {
-	    Schema::dropIfExists('articles');
-	    Schema::dropIfExists('article_translations');
-	    Schema::dropIfExists('article_tags');
+	    Schema::connection('dbp_users')->dropIfExists('articles');
+	    Schema::connection('dbp_users')->dropIfExists('article_translations');
+	    Schema::connection('dbp_users')->dropIfExists('article_tags');
     }
 }
